+++ src/app/actions/index.ts
@@ -1,5 +1,0 @@
-import * as LoginAction from './login.action';
-
-export type Actions =
-    LoginAction.Login |
-    LoginAction.LogOut;