+++ src/app/app.module.ts
@@ -1,64 +1,15 @@
-/**
- * Libraries
- */
 import { NgModule } from '@angular/core';
+
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpModule } from '@angular/http';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
 import { RouterModule } from '@angular/router';
-// import { StoreModule } from '@ngrx/store';
-// import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-// import { StoreLogMonitorModule, useLogMonitor } from '@ngrx/store-log-monitor';
 
-/**
- * Router
- */
-import { AppRoutingModule } from './app-routing.module';
+import { HttpModule } from '@angular/http';
 
-/**
- * Components
- * Containers
- * Models
- * Services
- * Stores
- */
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+
 import { AppComponent } from './app.component';
-import { COMPONENTS } from './components/index';
-import { CONTAINERS } from './containers/index';
-import { SERVICES } from './services/index';
-// import { STORES } from './stores/index';
 
-@NgModule({
-    declarations: [
-        AppComponent,
-        COMPONENTS,
-        CONTAINERS,
-    ],
-    imports: [
-        BrowserModule,
-        HttpModule,
-        ReactiveFormsModule,
-        FormsModule,
-        AppRoutingModule,
-        // StoreModule.forRoot({
-        //     STORES
-        // }),
-        // StoreDevtoolsModule.instrument({
-        //     monitor: useLogMonitor({
-        //         visible: false,
-        //         position: 'right'
-        //     })
-        // }),
-        // StoreLogMonitorModule
-    ],
-    providers: [
-        SERVICES
-    ],
-    bootstrap: [
-        AppComponent
-    ]
-})
-
-export class AppModule {
-
-}