+++ src/app/components/index.ts
@@ -1,5 +1,0 @@
-import { PageNotFoundComponent } from './page-not-found.component';
-
-export const COMPONENTS = [
-    PageNotFoundComponent
-];