+++ src/app/constants/ActionTypes.constant.ts
@@ -1,2 +1,0 @@
-export const LOG_IN: string = 'LOG_IN';
-export const LOG_OUT: string = 'LOG_OUT';