+++ src/app/containers/index.ts
@@ -1,5 +1,0 @@
-import * as LoginContainers from './login/index';
-
-export const CONTAINERS = [
-    LoginContainers.LoginComponent
-];