+++ src/app/containers/login/login.resource.ts
@@ -1,8 +1,0 @@
-export enum LoginResource {
-    placeholder_username = 'Username',
-    placeholder_password = 'Password',
-    text_remember_me = 'Remember Me',
-    text_sign_in = 'SIGN IN',
-    text_forgot_password = 'Forgot your password ?',
-    text_revolutionize = 'Revolutionzine your service orders'
-}