+++ src/app/models/user.model.ts
@@ -1,8 +1,0 @@
-export interface User {
-    userId: number;
-    userName: string;
-    password: string;
-    remember: boolean;
-}
-
-