+++ src/app/services/index.ts
@@ -1,5 +1,0 @@
-import { AuthenticationService } from './authentication.service';
-
-export const SERVICES = [
-    AuthenticationService
-];