+++ src/app/stores/index.ts
@@ -1,5 +1,0 @@
-import { LOGIN_ITEM } from './login.store';
-
-export const STORES = [
-    LOGIN_ITEM
-];